Implement Praos Protocol.
From the Babbage era, we intend to switch to using single-VRF Praos
(subsequently including Genesis) in place of TPraos, which has been in
use since Genesis. This commit introduces the Praos protocol in
preparation for use.

The main differences from TPraos are:
- Firstly, we move the full implementation of the protocol into this
  package, rather than splitting it between the ledger and network
  repositories.
- We have dropped support for the overlay which allowed some blocks to
  be issued under the round-robin BFT schedule. This is no longer needed
  since block production has been fully decentralised.
- We now include only a single VRF value, which is used to produce both
  the nonce and the leader election value.

Additional changes are directly crypto-related, and hence will be
included in a change of the crypto parameter associated with Babbage.