outbound-governor: check the state when promoting a warm peer
Check whether a peer is hasn't become cold when we try to update the
state once. In rare cases it could become cold just after we promoted it
to a hot peer, but before the completion continuation was scheduled.
This is because there could be a race between an asynchronous demotion
and the promotion.

Fixes #3550