peer-state-actions: remove no-op unregisterOutboundConnection
If a mini-protocol errored, or handshake errored, we don't need to
`unregisterOutboundConnection`.  The connection handler thrown an
exception (either mux does it, or in the other case the handler itself).
When this happens, the finally handler will govern the evolution of
state of the connection: it will set it to `TerminatingSt`, close the
connection and wait for the `TIME_WAIT` timer, when eventually the
connection will be forgotten.