Remove the use of stow/unstow ledger tables
The implementations of the 'StowableLedgerTables' functions call error,
which make sure we do not use them in the OnDisk tests of LedgerDB.

The ancillary fields of the ledger state that were needed for
implementing the 'StowableLedgerTables' instances are now removed.

The ledger DB implementation was modified to make sure we do not call
the 'StowableLedgerTables' methods if we request 'RunOnlyNew'.